Week 41 update

New:
-Corporation Assets Locations (x, y, z locations of assets)
-Corporation Assets Names (in-game names of assets)
-Corporation Bookmarks (and Bookmark folders)
-Corporation Contracts (and Contracts Items/Bids)
-Corporation Industry Jobs
-Corporation Market Orders
-Corporation Standings

Changed:
-BookmarkItem renamed to CharacterBookmarkItem (as we now have it for Corporations too)
-Character Notifications -> TypeEnum updated (no version change)
-Character and Corporation Transactions -> unitPrice changed from Integer to Float (no version change)
-Fixed universe/names test

+++ src/main/java/net/troja/eve/esi/model/CorporationWalletTransactionsResponse.java
@@ -51,7 +51,7 @@ public class CorporationWalletTransactionsResponse implements Serializable {
     private Integer typeId = null;
 
     @JsonProperty("unit_price")
-    private Integer unitPrice = null;
+    private Float unitPrice = null;
 
     public CorporationWalletTransactionsResponse clientId(Integer clientId) {
         this.clientId = clientId;
@@ -205,7 +205,7 @@ public class CorporationWalletTransactionsResponse implements Serializable {
         this.typeId = typeId;
     }
 
-    public CorporationWalletTransactionsResponse unitPrice(Integer unitPrice) {
+    public CorporationWalletTransactionsResponse unitPrice(Float unitPrice) {
         this.unitPrice = unitPrice;
         return this;
     }
@@ -216,11 +216,11 @@ public class CorporationWalletTransactionsResponse implements Serializable {
      * @return unitPrice
      **/
     @ApiModelProperty(example = "null", required = true, value = "Amount paid per unit")
-    public Integer getUnitPrice() {
+    public Float getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Integer unitPrice) {
+    public void setUnitPrice(Float unitPrice) {
         this.unitPrice = unitPrice;
     }
